requirements.txt
colorama==0.4.6
einops==0.8.0
et-xmlfile==1.1.0
filelock==3.13.4
fsspec==2024.3.1
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
lion-pytorch==0.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.4.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
